Add files via upload
Carbon Foot Print is the amount of carbon-di-oxide emitted through various human activities. As we all are aware of the climate change and global warming and carbon-di-oxide is the major contributor for it, we all are supposed to know our carbon foot prints and take necessary actions to reduce it.